pyproject.toml
[project]
name = "uv-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi[standard]>=0.115.8",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.2.4",
    "pydantic-settings>=2.7.1",
    "sqlmodel>=0.0.22",
    "torch==2.2.2",
]
